requirements.txt
# Core
httpx==0.27.0
aiohttp==3.9.3
pydantic==2.6.1
python-dotenv==1.0.1
loguru==0.7.2

# AI / LLM
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.6

# Data & Analysis
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
scipy==1.12.0

# Web Interface
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
rich==13.7.0

# Database
sqlalchemy==2.0.27
aiosqlite==0.19.0

# Scheduling
apscheduler==3.10.4

# Utilities
tenacity==8.2.3
loguru==0.7.2
cachetools==5.3.2
